Create ability to fetch samples remotely via urllib2
1. Samples must be hosted at HOST in certain directory
Using SimpleHTTPServer on localhost or 45.79.175.75. See code.

2. In remote or hybrid mode, they are fetched and loaded by
instrument. Fully supports multi-instrument changes if reloaded.

ISSUES:

Need to wrap the swmixer functions to byass seek and read that are
used with files but not streams. Ideally, create a new array/list for
each note that can be traversed or interated natively without the
underlying swmixer or wave methods.

Working on that next.